+++ .circleci/config.yml
@@ -9,7 +9,7 @@ jobs:
     docker:
       # specify the version you desire here
       # use `-browsers` prefix for selenium tests, e.g. `3.6.1-browsers`
-      - image: circleci/python:3.6.1
+      - image: circleci/3.7.0-stretch-node-browsers
       
       # Specify service dependencies here if necessary
       # CircleCI maintains a library of pre-built images
